chore: merge Root and routes files
