+++ app/renderer/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {createRoot} from 'react-dom/client';
 
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.jsx';
-import Root from './containers/Root.jsx';
+import Root from './Root.jsx';
 import store from './store.js';
 
 const container = document.getElementById('root');
@@ -15,8 +15,8 @@ root.render(
 );
 
 if (module.hot) {
-  module.hot.accept('./containers/Root', () => {
-    const NextRoot = require('./containers/Root.jsx').default;
+  module.hot.accept('./Root', () => {
+    const NextRoot = require('./Root.jsx').default;
     root.render(
       <ErrorBoundary>
         <NextRoot store={store} />
